introduce a general 'docker-hacks' role which also injects the systemctl replacement when running under the docker provider
