+++ .circleci/config.yml
@@ -16,11 +16,6 @@ jobs:
           name: Bring up Developer VM
           command: vagrant up --no-provision
       - run:
-          name: Inject docker systemctl replacement (required for testing in container only)
-          command: |
-            vagrant ssh -c 'sudo wget -O /usr/bin/systemctl https://raw.githubusercontent.com/gdraheim/docker-systemctl-replacement/master/files/docker/systemctl3.py'
-            vagrant ssh -c 'sudo mkdir -p /run/systemd/system/'
-      - run:
           name: Provision Developer VM
           command: UPDATE_VM_FLAGS=--provision-only vagrant provision
       - run:
