--- roles/x11-env/tasks/main.yml
@@ -1,9 +1,0 @@
----
-
-- name: Forward DISPLAY variable to the docker host
-  copy:
-    content: |
-      # forward the X11 DISPLAY to the docker host
-      export DISPLAY=host.docker.internal:0
-    dest: ~/.bashrc.d/x11-display.bash
-    mode: 0644
