+++ site.yml
@@ -14,7 +14,7 @@
     - { role: ansible-lint, tags: "ansible-lint" }
     - { role: testinfra, tags: "testinfra" }
     - { role: bashrc_d, tags: "bashrc_d" }
-    - { role: x11-env, tags: "x11-env", when: ansible_virtualization_type == "docker" }
+    - { role: docker-hacks, tags: "docker-hacks", when: ansible_virtualization_type == "docker" }
     - { role: cache, tags: "cache" }
     - { role: git, tags: "git" }
     - { role: vscode, tags: "vscode" }
